Rename components as lowercase
